Fix wrong data type and size used for bias data
Wrong data type or size is used for bias data. This results in bias
data being smaller than required and causes heap overflow. The bias
data is now declared with the correct data type and size.

+++ driver/support_library/tests/BranchingTests.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright © 2018-2021 Arm Limited.
+// Copyright © 2018-2022 Arm Limited.
 // SPDX-License-Identifier: Apache-2.0
 //
 
@@ -20,7 +20,7 @@ TEST_CASE("SimpleBranch")
     std::shared_ptr<Operand> input   = AddInput(network, TensorInfo({ 1, 16, 16, 16 })).tensor;
 
     std::shared_ptr<Constant> bias =
-        AddConstant(network, TensorInfo({ 1, 1, 1, 16 }, DataType::INT32_QUANTIZED), std::vector<uint8_t>(16, 0).data())
+        AddConstant(network, TensorInfo({ 1, 1, 1, 16 }, DataType::INT32_QUANTIZED), std::vector<int32_t>(16, 0).data())
             .tensor;
     std::shared_ptr<Constant> weights =
         AddConstant(network, TensorInfo({ 1, 1, 16, 16 }, DataType::UINT8_QUANTIZED, DataFormat::HWIO),
@@ -70,7 +70,7 @@ TEST_CASE("ReluAfterBranch")
     std::shared_ptr<Operand> input   = AddInput(network, TensorInfo({ 1, 16, 16, 16 })).tensor;
 
     std::shared_ptr<Constant> bias =
-        AddConstant(network, TensorInfo({ 1, 1, 1, 16 }, DataType::INT32_QUANTIZED), std::vector<uint8_t>(16, 0).data())
+        AddConstant(network, TensorInfo({ 1, 1, 1, 16 }, DataType::INT32_QUANTIZED), std::vector<int32_t>(16, 0).data())
             .tensor;
     std::shared_ptr<Constant> weights =
         AddConstant(network, TensorInfo({ 1, 1, 16, 16 }, DataType::UINT8_QUANTIZED, DataFormat::HWIO),
@@ -122,11 +122,11 @@ TEST_CASE("Branch in Sram")
 
     std::shared_ptr<Constant> bias1 =
         AddConstant(network, TensorInfo({ 1, 1, 1, 16 }, DataType::INT32_QUANTIZED, DataFormat::NHWC, { 0, 1.0f }),
-                    std::vector<uint8_t>(16, 0).data())
+                    std::vector<int32_t>(16, 0).data())
             .tensor;
     std::shared_ptr<Constant> bias2 =
         AddConstant(network, TensorInfo({ 1, 1, 1, 16 }, DataType::INT32_QUANTIZED, DataFormat::NHWC, { 0, 1.1f }),
-                    std::vector<uint8_t>(16, 0).data())
+                    std::vector<int32_t>(16, 0).data())
             .tensor;
     std::shared_ptr<Constant> weights =
         AddConstant(network, TensorInfo({ 1, 1, 16, 16 }, DataType::UINT8_QUANTIZED, DataFormat::HWIO, { 0, 1.f }),
